Completed draft of UI and moved fragments into their own package
