+++ app/src/main/java/de/hsaugsburg/teamulster/sohappy/fragment/CameraFragment.kt
@@ -1,4 +1,4 @@
-package de.hsaugsburg.teamulster.sohappy
+package de.hsaugsburg.teamulster.sohappy.fragment
 
 import android.Manifest
 import android.annotation.SuppressLint
@@ -14,6 +14,7 @@ import androidx.camera.core.ImageAnalysis
 import androidx.camera.lifecycle.ProcessCameraProvider
 import androidx.core.content.ContextCompat
 import androidx.databinding.DataBindingUtil
+import de.hsaugsburg.teamulster.sohappy.R
 import de.hsaugsburg.teamulster.sohappy.databinding.FragmentCameraBinding
 import de.hsaugsburg.teamulster.sohappy.util.YuvToRgbConverter
 import jp.co.cyberagent.android.gpuimage.GPUImage
