+++ app/src/main/java/de/hsaugsburg/teamulster/sohappy/fragment/JokesFragment.kt
@@ -1,9 +1,10 @@
-package de.hsaugsburg.teamulster.sohappy
+package de.hsaugsburg.teamulster.sohappy.fragment
 
 import android.os.Bundle
 import android.view.*
 import androidx.fragment.app.Fragment
 import androidx.databinding.DataBindingUtil
+import de.hsaugsburg.teamulster.sohappy.R
 import de.hsaugsburg.teamulster.sohappy.databinding.FragmentJokesBinding
 
 class JokesFragment : Fragment() {
