+++ app/src/main/java/de/hsaugsburg/teamulster/sohappy/fragment/Questionnaire04Fragment.kt
@@ -1,19 +1,18 @@
-package de.hsaugsburg.teamulster.sohappy
+package de.hsaugsburg.teamulster.sohappy.fragment
 
 import android.os.Bundle
-import android.util.Log
 import android.view.*
 import androidx.fragment.app.Fragment
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.activityViewModels
-import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
-import de.hsaugsburg.teamulster.sohappy.databinding.FragmentResultsBinding
+import de.hsaugsburg.teamulster.sohappy.R
+import de.hsaugsburg.teamulster.sohappy.databinding.FragmentQuestionnaire04Binding
 import de.hsaugsburg.teamulster.sohappy.viewmodel.QuestionnaireViewModel
 
-class ResultsFragment : Fragment() {
-    private lateinit var binding: FragmentResultsBinding
-    private val questionnaireViewModel: QuestionnaireViewModel by activityViewModels()
+class Questionnaire04Fragment: Fragment() {
+    private lateinit var binding: FragmentQuestionnaire04Binding
+    private val viewModel: QuestionnaireViewModel by activityViewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -22,13 +21,14 @@ class ResultsFragment : Fragment() {
     ): View? {
         binding = DataBindingUtil.inflate(
             inflater,
-            R.layout.fragment_results,
+            R.layout.fragment_questionnaire_04,
             container,
             false
         )
+        binding.viewModel = viewModel
 
-        binding.finishButton.setOnClickListener {
-            findNavController().navigate(R.id.homeFragment)
+        binding.continueButton.setOnClickListener {
+            findNavController().navigate(R.id.questionnaire05Fragment)
         }
 
         return binding.root
