+++ app/src/main/java/de/hsaugsburg/teamulster/sohappy/fragment/ResultsFragment.kt
@@ -1,18 +1,18 @@
-package de.hsaugsburg.teamulster.sohappy
+package de.hsaugsburg.teamulster.sohappy.fragment
 
 import android.os.Bundle
 import android.view.*
 import androidx.fragment.app.Fragment
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.activityViewModels
-import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
-import de.hsaugsburg.teamulster.sohappy.databinding.FragmentQuestionnaire04Binding
+import de.hsaugsburg.teamulster.sohappy.R
+import de.hsaugsburg.teamulster.sohappy.databinding.FragmentResultsBinding
 import de.hsaugsburg.teamulster.sohappy.viewmodel.QuestionnaireViewModel
 
-class Questionnaire04Fragment: Fragment() {
-    private lateinit var binding: FragmentQuestionnaire04Binding
-    private val viewModel: QuestionnaireViewModel by activityViewModels()
+class ResultsFragment : Fragment() {
+    private lateinit var binding: FragmentResultsBinding
+    private val questionnaireViewModel: QuestionnaireViewModel by activityViewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -21,14 +21,13 @@ class Questionnaire04Fragment: Fragment() {
     ): View? {
         binding = DataBindingUtil.inflate(
             inflater,
-            R.layout.fragment_questionnaire_04,
+            R.layout.fragment_results,
             container,
             false
         )
-        binding.viewModel = viewModel
 
-        binding.continueButton.setOnClickListener {
-            findNavController().navigate(R.id.questionnaire05Fragment)
+        binding.finishButton.setOnClickListener {
+            findNavController().navigate(R.id.homeFragment)
         }
 
         return binding.root
