+++ app/src/main/java/de/hsaugsburg/teamulster/sohappy/fragment/SmileFragment.kt
@@ -1,4 +1,4 @@
-package de.hsaugsburg.teamulster.sohappy
+package de.hsaugsburg.teamulster.sohappy.fragment
 
 import android.graphics.drawable.Animatable
 import android.os.Bundle
@@ -9,6 +9,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.databinding.DataBindingUtil
 import androidx.navigation.fragment.findNavController
+import de.hsaugsburg.teamulster.sohappy.R
 import de.hsaugsburg.teamulster.sohappy.databinding.FragmentSmileBinding
 
 // TODO: The requireView.postDelayed() calls serve as proof of concept for animations - replace!
