+++ app/src/main/res/layout/fragment_help.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".HelpFragment">
+    tools:context=".fragment.HelpFragment">
 
     <ScrollView
         android:layout_width="match_parent"
