+++ app/src/main/res/layout/fragment_smile.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".SmileFragment">
+    tools:context=".fragment.SmileFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
@@ -10,7 +10,7 @@
 
         <fragment
             android:id="@+id/gpuImageView"
-            android:name="de.hsaugsburg.teamulster.sohappy.CameraFragment"
+            android:name="de.hsaugsburg.teamulster.sohappy.fragment.CameraFragment"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             tools:layout="@layout/fragment_camera" />
@@ -65,7 +65,7 @@
             android:layout_width="100dp"
             android:layout_height="100dp"
             android:gravity="center"
-            android:text="3"
+            android:text="@string/fragment_camera_countdown_initial"
             android:textAppearance="@style/TextAppearanceCountdown"
             android:visibility="invisible"
             app:layout_constraintBottom_toBottomOf="parent"
