+++ app/src/main/res/navigation/navigation.xml
@@ -7,13 +7,13 @@
 
     <fragment
         android:id="@+id/settingsFragment"
-        android:name="de.hsaugsburg.teamulster.sohappy.SettingsFragment"
+        android:name="de.hsaugsburg.teamulster.sohappy.fragment.SettingsFragment"
         android:label="@string/fragment_settings_name"
         tools:layout="@layout/fragment_settings" />
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="de.hsaugsburg.teamulster.sohappy.HomeFragment"
+        android:name="de.hsaugsburg.teamulster.sohappy.fragment.HomeFragment"
         android:label="@string/app_name">
         <action
             android:id="@+id/action_homeFragment_to_settingsFragment"
@@ -31,12 +31,12 @@
 
     <fragment
         android:id="@+id/helpFragment"
-        android:name="de.hsaugsburg.teamulster.sohappy.HelpFragment"
+        android:name="de.hsaugsburg.teamulster.sohappy.fragment.HelpFragment"
         android:label="@string/fragment_help_name"
         tools:layout="@layout/fragment_help" />
     <fragment
         android:id="@+id/smileFragment"
-        android:name="de.hsaugsburg.teamulster.sohappy.SmileFragment"
+        android:name="de.hsaugsburg.teamulster.sohappy.fragment.SmileFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_smile">
         <action
@@ -45,7 +45,7 @@
     </fragment>
     <fragment
         android:id="@+id/questionnaire01Fragment"
-        android:name="de.hsaugsburg.teamulster.sohappy.Questionnaire01Fragment"
+        android:name="de.hsaugsburg.teamulster.sohappy.fragment.Questionnaire01Fragment"
         android:label="@string/fragment_questionnaire" >
         <action
             android:id="@+id/action_questionnaire01Fragment_to_questionnaire02Fragment"
@@ -53,7 +53,7 @@
     </fragment>
     <fragment
         android:id="@+id/questionnaire02Fragment"
-        android:name="de.hsaugsburg.teamulster.sohappy.Questionnaire02Fragment"
+        android:name="de.hsaugsburg.teamulster.sohappy.fragment.Questionnaire02Fragment"
         android:label="@string/fragment_questionnaire" >
         <action
             android:id="@+id/action_questionnaire02Fragment_to_questionnaire03Fragment"
@@ -61,7 +61,7 @@
     </fragment>
     <fragment
         android:id="@+id/questionnaire03Fragment"
-        android:name="de.hsaugsburg.teamulster.sohappy.Questionnaire03Fragment"
+        android:name="de.hsaugsburg.teamulster.sohappy.fragment.Questionnaire03Fragment"
         android:label="@string/fragment_questionnaire">
         <action
             android:id="@+id/action_questionnaire03Fragment_to_questionnaire04Fragment"
@@ -69,7 +69,7 @@
     </fragment>
     <fragment
         android:id="@+id/questionnaire04Fragment"
-        android:name="de.hsaugsburg.teamulster.sohappy.Questionnaire04Fragment"
+        android:name="de.hsaugsburg.teamulster.sohappy.fragment.Questionnaire04Fragment"
         android:label="@string/fragment_questionnaire" >
         <action
             android:id="@+id/action_questionnaire04Fragment_to_questionnaire05Fragment"
@@ -77,7 +77,7 @@
     </fragment>
     <fragment
         android:id="@+id/questionnaire05Fragment"
-        android:name="de.hsaugsburg.teamulster.sohappy.Questionnaire05Fragment"
+        android:name="de.hsaugsburg.teamulster.sohappy.fragment.Questionnaire05Fragment"
         android:label="@string/fragment_questionnaire">
         <action
             android:id="@+id/action_questionnaire05Fragment_to_questionnaire06Fragment"
@@ -85,7 +85,7 @@
     </fragment>
     <fragment
         android:id="@+id/questionnaire06Fragment"
-        android:name="de.hsaugsburg.teamulster.sohappy.Questionnaire06Fragment"
+        android:name="de.hsaugsburg.teamulster.sohappy.fragment.Questionnaire06Fragment"
         android:label="@string/fragment_questionnaire" >
         <action
             android:id="@+id/action_questionnaire06Fragment_to_resultsFragment"
@@ -93,7 +93,7 @@
     </fragment>
     <fragment
         android:id="@+id/resultsFragment"
-        android:name="de.hsaugsburg.teamulster.sohappy.ResultsFragment"
+        android:name="de.hsaugsburg.teamulster.sohappy.fragment.ResultsFragment"
         android:label="@string/fragment_results_appbar_title"
         tools:layout="@layout/fragment_results" >
         <action
@@ -102,7 +102,7 @@
     </fragment>
     <fragment
         android:id="@+id/jokesFragment"
-        android:name="de.hsaugsburg.teamulster.sohappy.JokesFragment"
+        android:name="de.hsaugsburg.teamulster.sohappy.fragment.JokesFragment"
         android:label="@string/fragment_jokes_appbar_title"
         tools:layout="@layout/fragment_jokes" />
