+++ app/src/main/res/values/styles.xml
@@ -44,6 +44,11 @@
         <item name="android:progressBackgroundTint">@android:color/darker_gray</item>
     </style>
 
+    <style name="SeekBarLabelAppearanceCustom">
+        <item name="android:textSize">18sp</item>
+        <item name="android:textColor">@android:color/darker_gray</item>
+    </style>
+
     <style name="SpinnerDropdownAppearanceCustom" parent="Widget.AppCompat.DropDownItem.Spinner">
         <item name="android:paddingStart">8dp</item>
     </style>
